pyproject.toml
[project]
name = "visitanalyze"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gql[all]>=4.0.0",
    "jsondiff>=2.2.1",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]
